requirements.txt
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
psycopg2-binary==2.9.10
requests>=2.26.0,<3.0